pyproject.toml
[project]
name = "tierkreis-fugaku"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "furo>=2025.7.19",
    "myst-nb>=1.3.0",
    "sphinx>=8.2.3",
    "sphinx-autodoc2>=0.5.0",
    "tierkreis>=2.0.2",
    "tkr-pytket-worker"
]

[tool.uv]
package=true

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
